refactor: Modularize Rust codebase to improve maintainability
This major refactoring addresses issue #1208 by breaking down the monolithic
1,372-line lib.rs file into a well-organized module structure:

## Module Structure Created:
- `error.rs`: Custom error types and CommandResult
- `state.rs`: Application state management
- `apm/`: APM analysis module with submodules:
  - `models.rs`: Data structures (APMStats, ToolUsage, etc.)
  - `analyzer.rs`: Core APM analysis logic
  - `historical.rs`: Historical data generation
  - `combined.rs`: CLI + SDK combination logic
  - `utils.rs`: Helper functions
- `commands/`: Tauri command modules:
  - `session.rs`: Session management commands
  - `apm.rs`: APM analysis commands
  - `history.rs`: History-related commands
  - `system.rs`: System utilities (greet, get_project_directory)

## Results:
- lib.rs reduced from 1,372 lines to 81 lines (94% reduction)
- Clear separation of concerns
- Improved code organization and maintainability
- All 57 existing tests pass
- Added 8 new tests for module accessibility

## Test Safety:
This refactoring was performed with the safety net of comprehensive tests
added in PR #1209, ensuring no functionality was broken during the process.
