+++ apps/desktop/src-tauri/src/commands/mod.rs
@@ -1,0 +1,6 @@
+//! Tauri command modules
+
+pub mod session;
+pub mod apm;
+pub mod history;
+pub mod system;
